pyproject.toml
[tool.poetry]
name = "release-logger"
version = "0.1.0"
description = ""
authors = ["Chien-De <cie001@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pymongo = "^4.6.1"
mongoengine = "^0.27.0"
flask = "^3.0.0"
flask-cors = "^4.0.0"
dnspython = "^2.4.2"
coverage = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
